+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.8.1 - In Progress
 
+- [FIX] New file manager uploading progress
 - [FIX] Fix build when no metrics enabled
 - [Feature] Single tap for infrared remotes
 # 1.8.0